reusing PricingData
