+++ unlock-app/src/components/interface/checkout/main/Confirm/ConfirmCard.tsx
@@ -1,7 +1,7 @@
 import { CheckoutService } from './../checkoutMachine'
 import { Connected } from '../../Connected'
 import { useConfig } from '~/utils/withConfig'
-import { Badge, Button, minifyAddress } from '@unlock-protocol/ui'
+import { Button } from '@unlock-protocol/ui'
 import { RiExternalLinkLine as ExternalLinkIcon } from 'react-icons/ri'
 import { Fragment, useState } from 'react'
 import { ToastHelper } from '~/components/helpers/toast.helper'
@@ -18,10 +18,11 @@ import { usePurchase } from '~/hooks/usePurchase'
 import { useUpdateUsersMetadata } from '~/hooks/useUserMetadata'
 import { usePricing } from '~/hooks/usePricing'
 import { usePurchaseData } from '~/hooks/usePurchaseData'
-import { formatNumber } from '~/utils/formatter'
 import { useFiatChargePrice } from '~/hooks/useFiatChargePrice'
 import { useCapturePayment } from '~/hooks/useCapturePayment'
 import { useCreditCardEnabled } from '~/hooks/useCreditCardEnabled'
+import { PricingData } from './PricingData'
+import { formatNumber } from '~/utils/formatter'
 
 interface Props {
   injectedProvider: unknown
@@ -33,13 +34,6 @@ interface CreditCardPricingBreakdownProps {
   total: number
   creditCardProcessingFee: number
   unlockServiceFee: number
-}
-
-interface PricingDataProps {
-  pricingData: any
-  lock: Lock
-  network: number
-  payment?: any
 }
 
 export function CreditCardPricingBreakdown({
@@ -94,58 +88,6 @@ export function CreditCardPricingBreakdown({
           </div>
         </div>
       </div>
-    </div>
-  )
-}
-
-export function PricingData({ pricingData, lock, payment }: PricingDataProps) {
-  return (
-    <div>
-      {!!pricingData?.prices?.length &&
-        pricingData.prices.map((item: any, index: number) => {
-          const first = index <= 0
-          const discount =
-            Number(lock!.keyPrice) > 0
-              ? (100 * (Number(lock!.keyPrice) - item.amount)) /
-                Number(lock!.keyPrice)
-              : 0
-          const symbol = payment?.route
-            ? payment.route.trade.inputAmount.currency.symbol
-            : item.symbol
-
-          return (
-            <div
-              key={index}
-              className={`flex border-b ${
-                first ? 'border-t' : null
-              } items-center justify-between text-sm px-0 py-2`}
-            >
-              <div>
-                1 Key for{' '}
-                <span className="font-medium">
-                  {minifyAddress(item.userAddress)}
-                </span>{' '}
-                {item.amount < Number(lock!.keyPrice) ? (
-                  <Badge variant="green" size="tiny">
-                    {discount}% Discount
-                  </Badge>
-                ) : null}
-              </div>
-
-              <div className="font-bold">
-                {item.amount <= 0
-                  ? 'FREE'
-                  : payment?.route
-                  ? `${formatNumber(
-                      payment.route
-                        .convertToQuoteToken(item.amount.toString())
-                        .toFixed()
-                    ).toLocaleString()} ${symbol}`
-                  : `${formatNumber(item.amount).toLocaleString()} ${symbol}`}
-              </div>
-            </div>
-          )
-        })}
     </div>
   )
 }
